Added rcopy.c and server.c
Rcopy is the client side process. It connects to a server and proceeds
to transfer files.
server.c is the server side process. It detects an attempted connection
from the client and starts a new process for them. Server maintains a
record of packets and writes them into the file in the proper order.

SendErr_init contains flags that can be turned on/off to enable or
disable packet drops, bit flips, and random seeds